requirements.txt
keras==2.4.3
scikit-learn==0.23.2
pillow==8.1.0
pandas==1.2.0
matplotlib==3.3.2
tqdm==4.55.1
theano
torch==1.7.1